읽어오는 코드 2
[문제]
읽어오는 코드에서 케이스 2에서 사이즈를 8 -> 4 바이트로 줄이니 num에서 읽어올때 id값을 읽어오는 것을 발견
[예상]
위에서부터 읽어오는 범위가 틀린건가?
[결론]
Player 클래스의 크기 : 64바이트
string 크기 : 32바이트
int 크기 : 4바이트
size_t 크기 : 8바이트
unique_ptr : 8바이트
인 것을 보아 score와 id 사이의 바이트 값을 읽어서 생기는 문제인 것으로 파악

+++ Homework/Homework.cpp
@@ -35,7 +35,7 @@ int main()
 			std::cout << "score : " << score << std::endl;
 			break;
 		case 2:
-			in.read((char*)&id, sizeof(size_t));
+			in.read((char*)&id, sizeof(int));
 			std::cout << "id : " << id << std::endl;
 			break;
 		case 3:
